무기 추가
Lazer 추가.
Lazer 대미지를 소숫점으로 설정했기 때문에 스크립트 일부 수정함.

+++ Space Shooter v2/Assets/BreakableAsteroids/Scripts/Fracture.cs
@@ -7,7 +7,7 @@ public class Fracture : Status
     [Tooltip("\"Fractured\" is the object that this will break into")]
     public GameObject fractured;
 
-    public override void GetDamage(int _damage)
+    public override void GetDamage(float _damage)
     {
         base.GetDamage(_damage);
